Create abstract class for X509CRLEntryStream with common DER functions
